backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2

# Security
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1

# Environment
python-dotenv==1.0.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# GitHub Integration
PyGithub==2.1.1

# AI/ML - Google Gemini
google-generativeai==0.3.2

# CORS (remove fastapi-cors as it's not a real package)
# Use FastAPI's built-in CORSMiddleware instead

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2


